// src/app/interfaces/libros.ts
  export interface Libro {
    id: number;
    title: string;
    authors: string;
    description: string;
    edition: string;
    format: string;
    num_pages: number;
    rating: number;
    rating_count: number;
    review_count: number;
    genres: string;
    genre_list: string;
    image_url: string;
    Quote1: string;
    Quote2: string;
    Quote3: string;
  }